pyproject.toml
[project]
name = "allegro-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.12.1",
    "faker>=40.11.0",
    "fastapi>=0.132.1",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "requests>=2.32.5",
    "types-requests>=2.32.4.20260107",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
typechecking = [
    'mypy',
    'pyright',
    'pyrefly',
    'pydantic-settings',
]


[tool.bandit]
exclude_dirs = ["tests"]

[tool.pytest.ini_options]
markers = [
    "num_of_offers: allows to specify number of initialized elements in db",
]
pythonpath = ["src"]
testpaths = ["tests"]

